requirements.txt
streamlit
tensorflow
scikit-learn
pillow
numpy
